RG CPU monitoring - beta

+++ FabricObserverTests/ObserverTests.cs
@@ -22,6 +22,7 @@ using System.Fabric.Health;
 using System.Fabric.Query;
 using System.IO;
 using System.Linq;
+using System.Runtime.CompilerServices;
 using System.Security.Cryptography;
 using System.Security.Cryptography.X509Certificates;
 using System.Threading;
@@ -96,6 +97,7 @@ namespace FabricObserverTests
             await DeployHealthMetricsAppAsync();
             await DeployTestApp42Async();
             await DeployVotingAppAsync();
+            await DeployCpuStressAppAsync();
         }
 
         [ClassCleanup]
@@ -426,6 +428,71 @@ namespace FabricObserverTests
             }
         }
 
+        private static async Task DeployCpuStressAppAsync()
+        {
+            string appName = "fabric:/CpuStress";
+
+            // If fabric:/Voting is already installed, exit.
+            var deployedTestApp =
+                    await FabricClient.QueryManager.GetDeployedApplicationListAsync(
+                            NodeName,
+                            new Uri(appName),
+                            TimeSpan.FromSeconds(30),
+                            Token);
+
+            if (deployedTestApp?.Count > 0)
+            {
+                return;
+            }
+
+            string appType = "CpuStressType";
+            string appVersion = "1.0.0";
+
+            // Change this to suit your configuration (so, if you are on Windows and you installed SF on a different drive, for example).
+            string imageStoreConnectionString = @"file:C:\SfDevCluster\Data\ImageStoreShare";
+            string packagePathInImageStore = "CpuStressApp";
+            string packagePathZip = Path.Combine(Environment.CurrentDirectory, "CpuStressApp.zip");
+            string packagePath = Path.Combine(Environment.CurrentDirectory, "CpuStressApp");
+
+            try
+            {
+                // Unzip the compressed HealthMetrics app package.
+                System.IO.Compression.ZipFile.ExtractToDirectory(packagePathZip, "CpuStressApp", true);
+
+                // Copy the HealthMetrics app package to a location in the image store.
+                FabricClient.ApplicationManager.CopyApplicationPackage(imageStoreConnectionString, packagePath, packagePathInImageStore);
+
+                // Provision the HealthMetrics application.          
+                await FabricClient.ApplicationManager.ProvisionApplicationAsync(packagePathInImageStore);
+
+                // Create HealthMetrics app instance.
+                ApplicationDescription appDesc = new(new Uri(appName), appType, appVersion);
+                await FabricClient.ApplicationManager.CreateApplicationAsync(appDesc);
+
+                // This is a hack. Withouth this timeout, the deployed test services may not have populated the FC cache?
+                // You may need to increase this value depending upon your dev machine? You'll find out..
+                await Task.Delay(TimeSpan.FromSeconds(15));
+            }
+            catch (FabricException fe)
+            {
+                if (fe.ErrorCode == FabricErrorCode.ApplicationAlreadyExists)
+                {
+                    await FabricClient.ApplicationManager.DeleteApplicationAsync(new DeleteApplicationDescription(new Uri(appName)) { ForceDelete = true });
+                    await DeployCpuStressAppAsync();
+                }
+                else if (fe.ErrorCode == FabricErrorCode.ApplicationTypeAlreadyExists)
+                {
+                    var appList = await FabricClient.QueryManager.GetApplicationListAsync(new Uri(appName));
+                    if (appList.Count > 0)
+                    {
+                        await FabricClient.ApplicationManager.DeleteApplicationAsync(new DeleteApplicationDescription(new Uri(appName)) { ForceDelete = true });
+                    }
+                    await FabricClient.ApplicationManager.UnprovisionApplicationAsync(appType, appVersion);
+                    await DeployCpuStressAppAsync();
+                }
+            }
+        }
+
         private static bool IsLocalSFRuntimePresent()
         {
             try
@@ -557,6 +624,33 @@ namespace FabricObserverTests
                 var deleteServiceDescription2 = new DeleteServiceDescription(new Uri(serviceName2));
                 await fabricClient.ServiceManager.DeleteServiceAsync(deleteServiceDescription1);
                 await fabricClient.ServiceManager.DeleteServiceAsync(deleteServiceDescription2);
+
+                // Delete an application instance from the application type.
+                var deleteApplicationDescription = new DeleteApplicationDescription(new Uri(appName));
+                await fabricClient.ApplicationManager.DeleteApplicationAsync(deleteApplicationDescription);
+
+                // Un-provision the application type.
+                await fabricClient.ApplicationManager.UnprovisionApplicationAsync(appType, appVersion);
+            }
+
+            // CpuStress \\
+
+            // Voting \\
+
+            if (await EnsureTestServicesExistAsync("fabric:/CpuStress"))
+            {
+                string appName = "fabric:/CpuStress";
+                string appType = "CpuStressType";
+                string appVersion = "1.0.0";
+                string serviceName = "fabric:/CpuStress/CpuMemStressor";
+                string packagePathInImageStore = "CpuStressApp";
+
+                // Clean up the unzipped directory.
+                fabricClient.ApplicationManager.RemoveApplicationPackage(imageStoreConnectionString, packagePathInImageStore);
+
+                // Delete services.
+                var deleteServiceDescription = new DeleteServiceDescription(new Uri(serviceName));
+                await fabricClient.ServiceManager.DeleteServiceAsync(deleteServiceDescription);
 
                 // Delete an application instance from the application type.
                 var deleteApplicationDescription = new DeleteApplicationDescription(new Uri(appName));
@@ -1114,7 +1208,7 @@ namespace FabricObserverTests
         // RG \\
 
         [TestMethod]
-        public async Task AppObserver_ObserveAsync_Successful_RGLimitWarningGenerated()
+        public async Task AppObserver_ObserveAsync_Successful_RGMemoryLimitWarningGenerated()
         {
             var startDateTime = DateTime.Now;
 
@@ -1125,7 +1219,34 @@ namespace FabricObserverTests
             using var obs = new AppObserver(TestServiceContext)
             {
                 MonitorDuration = TimeSpan.FromSeconds(1),
-                JsonConfigPath = Path.Combine(Environment.CurrentDirectory, "PackageRoot", "Config", "AppObserver_rg_warning.config.json"),
+                JsonConfigPath = Path.Combine(Environment.CurrentDirectory, "PackageRoot", "Config", "AppObserver_rg_memory_warning.config.json"),
+            };
+
+            await obs.ObserveAsync(Token);
+
+            // observer ran to completion with no errors.
+            Assert.IsTrue(obs.LastRunDateTime > startDateTime);
+
+            // observer detected warning conditions.
+            Assert.IsTrue(obs.HasActiveFabricErrorOrWarning);
+
+            // observer did not have any internal errors during run.
+            Assert.IsFalse(obs.IsUnhealthy);
+        }
+
+        [TestMethod]
+        public async Task AppObserver_ObserveAsync_Successful_RGCpuLimitWarningGenerated()
+        {
+            var startDateTime = DateTime.Now;
+
+            ObserverManager.FabricServiceContext = TestServiceContext;
+            ObserverManager.TelemetryEnabled = false;
+            ObserverManager.EtwEnabled = false;
+
+            using var obs = new AppObserver(TestServiceContext)
+            {
+                MonitorDuration = TimeSpan.FromSeconds(1),
+                JsonConfigPath = Path.Combine(Environment.CurrentDirectory, "PackageRoot", "Config", "AppObserver_rg_cpu_warning.config.json"),
             };
 
             await obs.ObserveAsync(Token);
@@ -1165,6 +1286,20 @@ namespace FabricObserverTests
                         TimeSpan.FromSeconds(60),
                         Token);
 
+            string svcManifest =
+                await FabricClient.ServiceManager.GetServiceManifestAsync(
+                        "VotingType", "1.0.0",
+                        "VotingWebPkg",
+                        TimeSpan.FromSeconds(60),
+                        Token);
+
+            string svcManifestData =
+                await FabricClient.ServiceManager.GetServiceManifestAsync(
+                        "VotingType", "1.0.0",
+                        "VotingDataPkg",
+                        TimeSpan.FromSeconds(60),
+                        Token);
+
             ApplicationList appList = await FabricClient.QueryManager.GetApplicationListAsync();
 
             ApplicationTypeList applicationTypeList =
@@ -1186,9 +1321,13 @@ namespace FabricObserverTests
             Assert.IsTrue(RGMemoryLimit == 2400);
 
             // A similar thing is tested here, without upgraded app parameters.
-            var (RGCpuEnabled, RGCpuLimit) = clientUtilities.TupleGetCpuResourceGovernanceInfo(appManifest, "VotingWebPkg", "Code", parameters);
+            var (RGCpuEnabled, RGCpuLimit) = clientUtilities.TupleGetCpuResourceGovernanceInfo(appManifest, svcManifest, "VotingWebPkg", "Code", parameters);
             Assert.IsTrue(RGCpuEnabled);
             Assert.IsTrue(RGCpuLimit == 3.5);
+
+            var (RGCpuEnabledData, RGCpuLimitData) = clientUtilities.TupleGetCpuResourceGovernanceInfo(appManifest, svcManifestData, "VotingDataPkg", "HelperExe2", parameters);
+            Assert.IsTrue(RGCpuEnabledData);
+            Assert.IsTrue(RGCpuLimitData == 0.6);
 
             // observer ran to completion with no errors.
             Assert.IsTrue(obs.LastRunDateTime > startDateTime);
@@ -2792,7 +2931,7 @@ namespace FabricObserverTests
                     Assert.IsTrue(data.RGMemoryEnabled && data.RGAppliedMemoryLimitMb > 0);     
                 }
 
-                if (data.ProcessName == "VotingData" || data.ProcessName == "VotingWeb")
+                if (data.ProcessName is "VotingData" or "VotingWeb" or "ConsoleApp6" or "ConsoleApp7")
                 {
                     Assert.IsTrue(data.RGCpuEnabled && data.RGAppliedCpuLimitCores > 0);
                 }
@@ -2858,7 +2997,7 @@ namespace FabricObserverTests
         public async Task AppObserver_ETW_RGMemoryLimitPercent_Warning()
         {
             using var foEtwListener = new FabricObserverEtwListener(_logger);
-            await AppObserver_ObserveAsync_Successful_RGLimitWarningGenerated();
+            await AppObserver_ObserveAsync_Successful_RGMemoryLimitWarningGenerated();
             List<ServiceTelemetryData> telemData = foEtwListener.foEtwConverter.ServiceTelemetryData;
 
             Assert.IsNotNull(telemData);
@@ -2869,6 +3008,24 @@ namespace FabricObserverTests
 
             // 2 service code packages + 2 helper code packages (VotingData) * 1 metric = 4 warnings...
             Assert.IsTrue(telemData.All(t => t.Metric == ErrorWarningProperty.RGMemoryUsagePercent && telemData.Count == 4));
+        }
+
+        // RG - warningRGCpuLimitPercent
+        [TestMethod]
+        public async Task AppObserver_ETW_RGCpuLimitPercent_Warning()
+        {
+            using var foEtwListener = new FabricObserverEtwListener(_logger);
+            await AppObserver_ObserveAsync_Successful_RGCpuLimitWarningGenerated();
+            List<ServiceTelemetryData> telemData = foEtwListener.foEtwConverter.ServiceTelemetryData;
+
+            Assert.IsNotNull(telemData);
+            Assert.IsTrue(telemData.Count > 0);
+
+            telemData = telemData.Where(
+                t => t.ApplicationName == "fabric:/CpuStress" && t.HealthState == HealthState.Warning).ToList();
+
+            // 2 service code packages + 2 helper code packages (VotingData) * 1 metric = 4 warnings...
+            Assert.IsTrue(telemData.All(t => t.Metric == ErrorWarningProperty.RGCpuUsagePercent && telemData.Count == 1));
         }
 
         // DiskObserver: TelemetryData \\
